new/backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pytest
httpx
pyyaml
